fix: noticket - remove shard_key from delete operations
... to ensure that only the _id index is used.

This should be a completely safe change, we are deleting by _id and
there is already a filter for the shard_key when finding the documents.

This is in place because we have observed mongo picking indexes other
than the _id index to delete documents by. This results in huge index
scans where a single batch of 500 docs takes 20 minutes to delete and
maxes out the read io on the primary.